+++ contracts/mock_contract/src/lib.rs
@@ -1,6 +1,8 @@
 mod custom_resp;
 mod msg_tests;
 
+use std::fmt::Debug;
+
 use cosmwasm_schema::{cw_serde, QueryResponses};
 use cosmwasm_std::{
     to_json_binary, Binary, Deps, DepsMut, Env, MessageInfo, Response, StdError, StdResult, Uint128,
@@ -55,7 +57,7 @@ where
 #[derive(cw_orch::QueryFns, QueryResponses)]
 pub enum QueryMsg<T = String>
 where
-    T: Serialize,
+    T: Serialize + Debug,
 {
     #[returns(String)]
     /// test-doc-comment
@@ -65,13 +67,19 @@ where
         /// test doc-comment
         t: T,
     },
-    #[returns(String)]
+    #[returns(ThirdReturn<T>)]
     ThirdQuery {
         /// test doc-comment
         t: T,
     },
     #[returns(u64)]
     FourthQuery(u64, String),
+}
+
+#[cw_serde]
+pub struct ThirdReturn<T> {
+    /// test doc-comment
+    pub t: T,
 }
 
 #[cw_serde]
@@ -148,7 +156,9 @@ pub fn query(_deps: Deps, _env: Env, msg: QueryMsg) -> StdResult<Binary> {
     match msg {
         QueryMsg::FirstQuery {} => to_json_binary("first query passed"),
         QueryMsg::SecondQuery { .. } => Err(StdError::generic_err("Query not available")),
-        QueryMsg::ThirdQuery { .. } => to_json_binary("third query passed"),
+        QueryMsg::ThirdQuery { .. } => to_json_binary(&ThirdReturn {
+            t: "third query passed",
+        }),
         QueryMsg::FourthQuery(_, _) => to_json_binary(&4u64),
     }
 }