+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## cw-orch-daemon 0.23.5
 
 - Fixed Get Tx By Events compatibility with Cosmos SDK 0.50+ for Daemon
+- Fix Generics on QueryMsg and Return types
 
 ## 0.23.0